+++ web/scripts/build-cli.js
@@ -14,7 +14,6 @@ const externalDependencies = [
   'web-push',
   'bonjour-service',
   'signal-exit',
-  'http-proxy-middleware',
   'multer',
   'mime-types',
 ];